+++ src/extensions/artcpaclub/src/esdt/_Unstaker.tsx
@@ -1,5 +1,6 @@
 import { sanitizeNumeric, shiftedBy } from '@peerme/core-ts'
 import { Button, Input } from '@peerme/web-ui'
+import { u64 } from '@vleap/warps'
 import React, { SyntheticEvent, useState } from 'react'
 import { useApp } from '../../../../shared/hooks/useApp'
 import { AppSection } from '../../../../shared/ui/elements'
@@ -24,7 +25,7 @@ export function _Unstaker(props: Props) {
       Contracts(app.config).EsdtUserUnstake.Address,
       Contracts(app.config).EsdtUserUnstake.Endpoint,
       0n,
-      [props.pool.pool_id, amountBig],
+      [u64(BigInt(props.pool.pool_id)), u64(amountBig)],
       []
     )
   }
